complete remaining tasks from Antigravity work plan
- Fixed telemetry collector encapsulation (C2) with get_records() method
- Added bootstrap services initialization in CLI (D1, D2) for checkpoint manager and telemetry
- Created integration tests for distributed execution stress testing (E1, E2, E3)
- Verified all existing functionality remains intact

+++ src/em_cubed/surfaces/wasm_surface.py
@@ -186,7 +186,7 @@ class WASMSurface(SurfaceBase):
                 elif "i64" in type_str:
                     args.append(int(val))
                 elif "f32" in type_str or "f64" in type_str:
-                    args.append(int(val))
+                    args.append(float(val))
                 else:
                     args.append(val)
             except Exception:
